feature/backend: Reusable code and refactored services
